+++ src/Libraries/SmartStore.Data/Migrations/MigrationsConfiguration.cs
@@ -43,8 +43,29 @@
 		public void MigrateLocaleResources(LocaleResourcesBuilder builder)
 		{
 			builder.AddOrUpdate("Admin.Configuration.Settings.ShoppingCart.ThirdPartyEmailHandOver.Hint",
-				"Specifies whether customers can agree to a transferring of their email address to third parties when ordering, and whether the checkbox is enabled by default during checkout. Please keep in mind that the option 'Show activated' isn't legally justified in line with the GDPR.",
+				"Specifies whether customers can agree to a transferring of their email address to third parties when ordering, and whether the checkbox is enabled by default during checkout. Please note that the 'Show activated' option isn't legally compliant in line with the GDPR.",
 				"Legt fest, ob Kunden bei einer Bestellung der Weitergabe ihrer E-Mail Adresse an Dritte zustimmen können und ob die Checkbox dafür standardmäßig aktiviert ist. Bitte beachten Sie, dass die Option 'Aktiviert anzeigen' im Rahmen der DSVGO nicht rechtskonform ist.");
+
+			builder.AddOrUpdate("Admin.Configuration.Settings.CustomerUser.Privacy", "Privacy", "Datenschutz");
+			
+			builder.AddOrUpdate("Admin.Configuration.Settings.CustomerUser.Privacy.EnableCookieConsent",
+				"Enable cookie consent", 
+				"Cookie-Hinweis aktivieren",
+				"Specifies whether the cookie consent box will be displayed in the frontend.", 
+				"Legt fest, ob ein Element für die Zustimmung zur Nutzung von Cookies im Frontend angezeigt wird.");
+
+			builder.AddOrUpdate("Admin.Configuration.Settings.CustomerUser.Privacy.CookieConsentBadgetext",
+				"Cookie consent display text",
+				"Cookie-Hinweistext",
+				"Specifies the text, that will be displayed to your customers if they havn't agreed to the usage of cookis yet.",
+				"Bestimmt den Text, der Ihren Kunden beim Besuch der Seite angezeigt wird, sofern Sie ihre Zustimmung zur Nutzung von Cookies noch nicht gegeben haben.");
+			
+			builder.AddOrUpdate("CookieConsent.BadgeText",
+				"{0} is using cookies, to guarantee to best shopping experience. Partially cookies will be set by thrid parties. <a href='{1}'>Privacy Info</a>",
+				"{0} benutzt Cookies, um Ihnen das beste Einkaufs-Erlebnis zu ermöglichen. Zum Teil werden Cookies auch von Drittanbietern gesetzt. <a href='{1}'>Datenschutzerklärung</a>");
+
+			builder.AddOrUpdate("CookieConsent.Button", "Okay, got it", "Ok, verstanden");
+			
 		}
 	}
 }