+++ src/Presentation/SmartStore.Web.Framework/DependencyRegistrar.cs
@@ -691,6 +691,7 @@ namespace SmartStore.Web.Framework
 			{
 				pageHelperRegistration.PropertiesAutowired(PropertyWiringOptions.None);
 				builder.RegisterType<HandleExceptionFilter>().AsActionFilterFor<SmartController>(-100);
+				builder.RegisterType<CookieConsentFilter>().AsActionFilterFor<PublicControllerBase>().InstancePerRequest();
 			}
 		}