+++ src/Presentation/SmartStore.Web/Controllers/CommonController.cs
@@ -41,6 +41,9 @@ using SmartStore.Web.Framework.Theming;
 using SmartStore.Web.Framework.UI;
 using SmartStore.Web.Infrastructure.Cache;
 using SmartStore.Web.Models.Common;
+using SmartStore.Web.Framework.Filters;
+using SmartStore.Core.Plugins;
+using SmartStore.Core.Caching;
 
 namespace SmartStore.Web.Controllers
 {
@@ -61,7 +64,8 @@ namespace SmartStore.Web.Controllers
 
 		private readonly StoreInformationSettings _storeInfoSettings;
 		private readonly CustomerSettings _customerSettings;
-        private readonly TaxSettings _taxSettings;
+		private readonly PrivacySettings _privacySettings;
+		private readonly TaxSettings _taxSettings;
         private readonly CatalogSettings _catalogSettings;
         private readonly ShoppingCartSettings _shoppingCartSettings;
         private readonly ThemeSettings _themeSettings;
@@ -83,7 +87,7 @@ namespace SmartStore.Web.Controllers
         private readonly Lazy<IShoppingCartService> _shoppingCartService;
 
 		private readonly IBreadcrumb _breadcrumb;
-
+		
 		public CommonController(
 			ICommonServices services,
 			ITopicService topicService,
@@ -97,8 +101,9 @@ namespace SmartStore.Web.Controllers
 			Lazy<ICompareProductsService> compareProductsService,
 			Lazy<IUrlRecordService> urlRecordService,
 			StoreInformationSettings storeInfoSettings,
-            CustomerSettings customerSettings, 
-            TaxSettings taxSettings, 
+            CustomerSettings customerSettings,
+			PrivacySettings privacySettings,
+			TaxSettings taxSettings, 
 			CatalogSettings catalogSettings,
             ShoppingCartSettings shoppingCartSettings,
             EmailAccountSettings emailAccountSettings,
@@ -133,7 +138,8 @@ namespace SmartStore.Web.Controllers
 
 			_storeInfoSettings = storeInfoSettings;
 			_customerSettings = customerSettings;
-            _taxSettings = taxSettings;
+			_privacySettings = privacySettings;
+			_taxSettings = taxSettings;
             _catalogSettings = catalogSettings;
             _shoppingCartSettings = shoppingCartSettings;
             _commonSettings = commonSettings;
@@ -156,7 +162,7 @@ namespace SmartStore.Web.Controllers
             _shoppingCartService = shoppingCartService;
 
 			_breadcrumb = breadcrumb;
-        }
+		}
 
         #region Utilities
 
@@ -843,6 +849,46 @@ namespace SmartStore.Web.Controllers
 
 				return model;			
 			}, TimeSpan.FromMinutes(1) /* 1 min. (just for the duration of pdf processing) */);
+		}
+
+		[ChildActionOnly]
+		public ActionResult CookieConsentBadge()
+		{
+			if (!_privacySettings.EnableCookieConsent)
+			{
+				return new EmptyResult();
+			}
+			
+			var model = new CookieConsentModel();
+
+			if (!_privacySettings.CookieConsentBadgetext.HasValue())
+			{
+				// loads default value if it's empty (must be done this way as localized values can't be initial values of settings)
+				model.BadgeText = T("CookieConsent.BadgeText", 
+					_services.StoreContext.CurrentStore.Name, 
+					Url.RouteUrl("Topic", new { SeName = Url.TopicSeName("PrivacyInfo") }));
+			}
+			else
+			{
+				model.BadgeText = _privacySettings.GetLocalized(x => x.CookieConsentBadgetext).Value.FormatWith(
+					_services.StoreContext.CurrentStore.Name,
+					Url.RouteUrl("Topic", new { SystemName = "PrivacyInfo" })
+				);
+			}
+			
+			var consentCookie = this.Request.Cookies[CookieConsent.CONSENT_COOKIE_NAME];
+			if (consentCookie != null && consentCookie.Value == "true")
+				return new EmptyResult();
+
+			return PartialView(model);
+		}
+
+		[HttpPost]
+		public ActionResult CookieConsentBadge(CookieConsentModel model)
+		{
+			CookieConsent.SetCookieConsent(Response, true);
+
+			return new EmptyResult();
 		}
 
 		#endregion