+++ src/Presentation/SmartStore.Web/Models/Common/CookieConsentModel.cs
@@ -1,0 +1,9 @@
+﻿using SmartStore.Web.Framework.Modelling;
+
+namespace SmartStore.Web.Models.Common
+{
+    public partial class CookieConsentModel : ModelBase
+    {
+		public string BadgeText { get; set; }
+	}
+}