+++ changelog.md
@@ -1,10 +1,11 @@
-﻿# Release Notes
+# Release Notes
 
 ## SmartStore.NET 3.1.5
 ### New Features
 * #1429 Search engine friendly topic URLs
 * Trusted Shops: Added consent checkbox to confirm order page for submission of customer email address to Trusted Shops if review widget is active
 * #1226 Shop-Connector: Added exchange of tier prices and delivery times
+* Implemented cookie consent according to EU law
 
 ### Improvements
 * Added double opt-in feature for newsletter subscriptions during checkout (confirm order)